Implement class Robot
It's a simple getter / setter for the robot position and direction
(x, y, face). Thus it simply implement method place and report.

The other methods like move, right, left will be implement by another
class. This other class will ensure that the robot does not leave the
table ... what is not the responsibility of the robot.